+++ GetClosePrice/Models/PriceViewModel.cs
@@ -16,7 +16,7 @@ namespace GetClosePrice
                 this.Name = item[1];
                 this.Volumn = (int)float.Parse(item[2]);
                 this.Count = (int)float.Parse(item[3]);
-                this.Amount = (int)float.Parse(item[4]);
+                this.Amount = (long)float.Parse(item[4]);
                 this.Open = IsNumeric(item[5]) ? double.Parse(item[5]) : 0;
                 this.High = IsNumeric(item[6]) ? double.Parse(item[6]) : 0;
                 this.Low = IsNumeric(item[7]) ? double.Parse(item[7]) : 0;
@@ -27,7 +27,7 @@ namespace GetClosePrice
                 this.Code = int.Parse(item[0]);
                 this.Name = item[1];
                 this.Count = (int)float.Parse(item[10]);
-                this.Amount = (int)float.Parse(item[9]);
+                this.Amount = (long)float.Parse(item[9]);
                 this.Volumn = (int)float.Parse(item[8]);
                 this.Open = IsNumeric(item[4]) ? double.Parse(item[4]) : 0;
                 this.High = IsNumeric(item[5]) ? double.Parse(item[5]) : 0;
@@ -43,7 +43,7 @@ namespace GetClosePrice
         public double Low { get; set; }
         public int Volumn { get; set; }
         public int Count { get; set; }
-        public int Amount { get; set; }
+        public long Amount { get; set; }
 
         private static bool IsNumeric(string s)
         {
